pyproject.toml
[project]
name = "guarda-din-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
